+++ packages/expo-brightness/src/Brightness.ts
@@ -1,9 +1,5 @@
-import {
-  createPermissionHook,
-  type EventSubscription,
-  type PermissionResponse,
-  UnavailabilityError,
-} from 'expo-modules-core';
+import { type PermissionResponse, createPermissionHook } from 'expo';
+import { type EventSubscription, UnavailabilityError } from 'expo-modules-core';
 import { Platform } from 'react-native';
 
 import ExpoBrightness from './ExpoBrightness';