+++ packages/expo-camera/src/index.ts
@@ -1,5 +1,5 @@
+import { createPermissionHook } from 'expo';
 import type { PermissionResponse } from 'expo-modules-core';
-import { createPermissionHook } from 'expo-modules-core';
 
 import type { BarcodeScanningResult, BarcodeType } from './Camera.types';
 import CameraManager from './ExpoCameraManager';